CI | Add multi-arch support for Linux (amd64, arm64, armhf) and macOS universal binary
- Add .cargo/config.toml with cross-linker config for ARM targets
- Update Dockerfile with ARM cross-compilers, multi-arch dev libraries,
  QEMU user-mode emulation, and Rust cross-compilation targets
- Update CI workflow with build matrix for 3 Linux architectures and a
  macOS universal binary job (lipo aarch64 + x86_64)
- Add multi-arch packaging (.deb/.rpm) for all Linux targets
- Update release job to publish arch-specific Linux binaries/packages
  and a single macOS universal binary
- Bump CI image tag to 2.0, add :latest alias
- Update pclsync submodule with aarch64/LP64 platform detection fixes

+++ src/wrapper/crypto.rs
@@ -288,7 +288,7 @@ impl PCloudClient {
     /// Unix timestamp of when crypto expires, or 0 if never set up.
     pub fn crypto_expires(&self) -> i64 {
         // Safety: psync_crypto_expires is safe to call anytime
-        unsafe { raw::psync_crypto_expires() }
+        unsafe { raw::psync_crypto_expires() as i64 }
     }
 
     /// Get the crypto password hint.
